[luci/service] Migrate StridedSlice shape inference rule
This migrates StridedSlice shape inference rule to sinf::Algorithm.

ONE-DCO-1.0-Signed-off-by: sunki <ajsthfldu@outlook.com>
